+++ sme/src/sme_model.cpp
@@ -147,12 +147,13 @@ void pybindModel(pybind11::module &m) {
            Args:
                filename (str): the name of the geometry image to import
            )")
-      .def("simulate", &sme::Model::simulateFloat, pybind11::arg("simulation_time"),
-           pybind11::arg("image_interval"),
+      .def("simulate", &sme::Model::simulateFloat,
+           pybind11::arg("simulation_time"), pybind11::arg("image_interval"),
            pybind11::arg("timeout_seconds") = 86400,
            pybind11::arg("throw_on_timeout") = true,
            pybind11::arg("simulator_type") = simulate::SimulatorType::Pixel,
            pybind11::arg("continue_existing_simulation") = false,
+           pybind11::arg("return_results") = true,
            R"(
            returns the results of the simulation.
 
@@ -160,9 +161,10 @@ void pybindModel(pybind11::module &m) {
                simulation_time (float): The length of the simulation in model units of time, e.g. `5.5`
                image_interval (float): The interval between images in model units of time, e.g. `1.1`
                timeout_seconds (int): The maximum time in seconds that the simulation can run for. Default value: 86400 = 1 day.
-               throw_on_timeout (bool): Whether to throw an exception on simulation timeout. Default value: `true`.
+               throw_on_timeout (bool): Whether to throw an exception on simulation timeout. Default value: `True`.
                simulator_type (sme.SimulatorType): The simulator to use: `sme.SimulatorType.DUNE` or `sme.SimulatorType.Pixel`. Default value: Pixel.
-               continue_existing_simulation (bool): Whether to continue the existing simulation, or start a new simulation. Default value: `false`, i.e. any existing simulation results are discarded before doing the simulation.
+               continue_existing_simulation (bool): Whether to continue the existing simulation, or start a new simulation. Default value: `False`, i.e. any existing simulation results are discarded before doing the simulation.
+               return_results (bool): Whether to return the simulation results. Default value: `True`. If `False`, an empty SimulationResultList is returned.
 
            Returns:
                SimulationResultList: the results of the simulation
@@ -170,12 +172,13 @@ void pybindModel(pybind11::module &m) {
            Raises:
                RuntimeError: if the simulation times out or fails
            )")
-      .def("simulate", &sme::Model::simulateString, pybind11::arg("simulation_times"),
-           pybind11::arg("image_intervals"),
+      .def("simulate", &sme::Model::simulateString,
+           pybind11::arg("simulation_times"), pybind11::arg("image_intervals"),
            pybind11::arg("timeout_seconds") = 86400,
            pybind11::arg("throw_on_timeout") = true,
            pybind11::arg("simulator_type") = simulate::SimulatorType::Pixel,
            pybind11::arg("continue_existing_simulation") = false,
+           pybind11::arg("return_results") = true,
            R"(
            returns the results of the simulation.
 
@@ -186,29 +189,31 @@ void pybindModel(pybind11::module &m) {
                throw_on_timeout (bool): Whether to throw an exception on simulation timeout. Default value: `true`.
                simulator_type (sme.SimulatorType): The simulator to use: `sme.SimulatorType.DUNE` or `sme.SimulatorType.Pixel`. Default value: Pixel.
                continue_existing_simulation (bool): Whether to continue the existing simulation, or start a new simulation. Default value: `false`, i.e. any existing simulation results are discarded before doing the simulation.
+               return_results (bool): Whether to return the simulation results. Default value: `True`. If `False`, an empty SimulationResultList is returned.
 
            Returns:
                SimulationResultList: the results of the simulation
 
            Raises:
                RuntimeError: if the simulation times out or fails
-           )")      .def("__repr__",
+           )")
+      .def("__repr__",
            [](const sme::Model &a) {
              return fmt::format("<sme.Model named '{}'>", a.getName());
            })
       .def("__str__", &sme::Model::getStr);
 }
 
-static std::vector<SimulationResult> getSimulationResults(const simulate::Simulation *sim) {
+static std::vector<SimulationResult>
+getSimulationResults(const simulate::Simulation *sim) {
   std::vector<SimulationResult> results;
   for (std::size_t i = 0; i < sim->getTimePoints().size(); ++i) {
-    auto& result = results.emplace_back();
+    auto &result = results.emplace_back();
     result.timePoint = sim->getTimePoints()[i];
     result.concentrationImage = toPyImageRgb(sim->getConcImage(i, {}, true));
     std::tie(result.speciesConcentration, result.speciesDcdt) =
         sim->getPyConcs(i);
   }
-
   return results;
 }
 
@@ -235,7 +240,6 @@ void Model::init() {
     parameters.emplace_back(s.get(), paramId.toStdString());
   }
   compartmentImage = toPyImageRgb(s->getGeometry().getImage());
-
 }
 
 Model::Model(const std::string &filename) {
@@ -293,10 +297,11 @@ void Model::exportSmeFile(const std::string &filename) {
   s->exportSMEFile(filename);
 }
 
-std::vector<SimulationResult> Model::simulateString(const std::string &lengths, const std::string &intervals,
-                     int timeoutSeconds, bool throwOnTimeout,
-                     simulate::SimulatorType simulatorType,
-                     bool continueExistingSimulation) {
+std::vector<SimulationResult>
+Model::simulateString(const std::string &lengths, const std::string &intervals,
+                      int timeoutSeconds, bool throwOnTimeout,
+                      simulate::SimulatorType simulatorType,
+                      bool continueExistingSimulation, bool returnResults) {
   QElapsedTimer simulationRuntimeTimer;
   simulationRuntimeTimer.start();
   double timeoutMillisecs{static_cast<double>(timeoutSeconds) * 1000.0};
@@ -319,17 +324,21 @@ std::vector<SimulationResult> Model::simulateString(const std::string &lengths, 
   if (const auto &e = sim->errorMessage(); throwOnTimeout && !e.empty()) {
     throw SmeRuntimeError(fmt::format("Error during simulation: {}", e));
   }
-  return getSimulationResults(sim.get());
+  if(returnResults) {
+    return getSimulationResults(sim.get());
+  }
+  return {};
 }
 
-std::vector<SimulationResult> Model::simulateFloat(double simulationTime, double imageInterval,
+std::vector<SimulationResult>
+Model::simulateFloat(double simulationTime, double imageInterval,
                      int timeoutSeconds, bool throwOnTimeout,
                      simulate::SimulatorType simulatorType,
-                     bool continueExistingSimulation) {
+                     bool continueExistingSimulation, bool returnResults) {
   return simulateString(QString::number(simulationTime, 'g', 17).toStdString(),
-                  QString::number(imageInterval, 'g', 17).toStdString(),
-                  timeoutSeconds, throwOnTimeout, simulatorType,
-                  continueExistingSimulation);
+                        QString::number(imageInterval, 'g', 17).toStdString(),
+                        timeoutSeconds, throwOnTimeout, simulatorType,
+                        continueExistingSimulation, returnResults);
 }
 
 std::string Model::getStr() const {